fix call to json_array_contains_value passing integer when a string key is required.,
For JSON types, keys are strings. The previous implementation passed an integer type. This may work in mariadb and postgres, but it does not work in mysql. It yields:

```
(mysql.connector.errors.DataError) 3146 (22032): Invalid data type for JSON data in argument 1 to function json_contains; a JSON string or JSON type is required.
```

I've tested this on my personal Unraid server running Romm on Docker with MySQL.

This may be better implemented by modifying the implementation of `json_array_contains_value`, but I am unsure if there are other impacts and this is the only affected case I have seen.